Move new classes to src/component directory, per discussion
The namespace root is currently src/.
So src/component/class-carousel.php
will have the class Amp\AmpWP\Component\Carousel.
