+++ composer.json
@@ -48,7 +48,7 @@
   },
   "autoload": {
     "psr-4": {
-      "AMP\\": "includes/amp/"
+      "Amp\\AmpWP\\": "src/"
     }
   }
 }
