+++ src/component/class-image.php
@@ -5,7 +5,7 @@
  * @package AMP
  */
 
-namespace AMP;
+namespace Amp\AmpWP\Component;
 
 /**
  * Class Image
